api/api_requirements.txt
Flask>=3.1.1
python-dateutil>=2.9.0.post0
gunicorn>=21.2
requests>=2.32.0
